@@ -1,23 +1,24 @@
-name: lab_sound_flutter
-description: "LabSound lib for flutter"
+name: tau_labs
+description: "LabSound library for Flutter"
 version: 0.0.1
 homepage:
 
 environment:
-  sdk: ^3.5.1
+  sdk: ^3.7.2
   flutter: '>=3.3.0'
 
 dependencies:
   flutter:
     sdk: flutter
   plugin_platform_interface: ^2.0.2
+  rxdart: ^0.28.0
 
 dev_dependencies:
-  ffi: ^2.1.0
-  ffigen: ^11.0.0
+  ffi: ^2.1.3
+  ffigen: ^13.0.0
   flutter_test:
     sdk: flutter
-  flutter_lints: ^4.0.0
+  flutter_lints: ^5.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
