@@ -2,12 +2,12 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:lab_sound_flutter/lab_sound_flutter.dart';
+import 'package:tau_labs/tau_labs.dart';
 
 class DrawFrequency extends StatefulWidget {
   final AnalyserNode analyserNode;
 
-  DrawFrequency(this.analyserNode, {Key? key}) : super(key: key);
+  const DrawFrequency(this.analyserNode, {super.key});
 
   @override
   _DrawFrequencyState createState() => _DrawFrequencyState();
@@ -58,10 +58,10 @@ class DrawFrequencyPainter extends CustomPainter {
       ..strokeWidth = 1
       ..style = PaintingStyle.fill
       ..strokeCap = StrokeCap.round;
-    double w = size.width / this.analyserNode.frequencyBinCount;
+    double w = size.width / analyserNode.frequencyBinCount;
     double h = size.height / 256.0;
     int n = 0;
-    this.analyserNode.getByteFrequencyData().forEach((val) {
+    analyserNode.getByteFrequencyData().forEach((val) {
       canvas.drawRect(
           Rect.fromLTWH(n * w, size.height - val * h,
               max(w.ceil().toDouble(), 1.0), val * h),
